Demo changes
SongVal is now hard-coded. This will end up being whatever song we
choose to demo.
Clef is now set appropriately when "process" is hit.
Added new "PlayPage" activity that will handle the MediaPlayer. This
page includes play/pause/stop buttons, as well as buttons to go back the
properties page OR the MainActivity