Key signatures, new song
Key signatures definitely work for treble clef but something is wrong
with bass clef and I don't know what it is.
Also the app now plays "Arabesque" instead of "Ode to Joy," and the
former includes SIXTEENTH NOTES. Which our app doesn't even support but
I forced it to. yay.